Refactor plugin for more language support
