+++ src/main/resources/META-INF/gradle-plugins/com.google.protobuf.java.properties
@@ -1,0 +1,1 @@
+implementation-class=com.google.protobuf.gradle.plugins.ProtobufJavaPlugin
