+++ src/main/resources/META-INF/gradle-plugins/com.google.protobuf.properties
@@ -1,1 +1,1 @@
-implementation-class=com.google.protobuf.gradle.ProtobufPlugin
+implementation-class=com.google.protobuf.gradle.plugins.ProtobufPlugin
